Backend/requirements.txt
fastapi
fastapi[standard]
faiss-cpu
numpy
transformers
torch
uvicorn